Added baseline matrices for IBM regression tests
While refactoring the parallel methods, several issues were found and
fixed, resulting in changed values in the matrices so new matrices had
to be generated. The original matrices are dropped from the git history
to save space. The new matrices are also smaller than the original
matrices used during the parallel IBM development.